pyproject.toml
[project]
authors = [{ name = "de-odex", email = "de-odex@users.noreply.github.com" }]
name = "mygo_serifu_bot"
requires-python = ">= 3.11"
version = "0.1.0"
dependencies = [
  "discord>=2.3.2,<3",
  "python-dotenv>=1.0.1,<2",
  "typed-ffmpeg>=2.6.2,<3",
  "requests>=2.32.3,<3",
  "loguru>=0.7.3,<0.8",
  "pysubs2>=1.8.0,<2",
  "whoosh>=2.7.4,<3",
  "watchfiles>=1.0.4,<2",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["win-64", "linux-64"]

[tool.pixi.pypi-dependencies]
mygo_serifu_bot = { path = ".", editable = true }

[tool.pixi.tasks]
start = "python -m mygo_serifu_bot.main"
